added token check to all requests

+++ backend/src/main/java/de/damien/funnyplaces/comments/CommentService.java
@@ -1,5 +1,6 @@
 package de.damien.funnyplaces.comments;
 
+import de.damien.funnyplaces.accounts.AccountService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -18,35 +19,42 @@ public class CommentService {
     }
 
     public Comment addComment(Comment comment, String token) throws AuthenticationException {
-        //TODO uncomment later
-//        String userByToken = AccountService.getAccountByToken(token);
-//        if (!AccountService.authenticateUser(comment.getWriter().getName(), token)) {
-//            //token authentication failed
-//            throw new AuthenticationException("");
-//        }
+        if (!AccountService.authenticateUser(comment.getWriter().getName(), token)) {
+            throw new AuthenticationException("Invalid token");
+        }
         return commentRepository.save(comment);
     }
 
-    public Comment getComment(Long id) throws NoSuchElementException {
+    public Comment getComment(Long id, String token) throws NoSuchElementException, AuthenticationException {
         Optional<Comment> commentOptional = commentRepository.findById(id);
         if (commentOptional.isEmpty()) {
             throw new NoSuchElementException();
         }
-        return commentOptional.get();
+        Comment comment = commentOptional.get();
+        if (AccountService.getAccountByToken(token) == null) {
+            throw new AuthenticationException("Invalid token");
+        }
+        return comment;
     }
 
-    public Long updateComment(Long id, Comment commentNew) throws NoSuchElementException {
+    public Long updateComment(Long id, Comment commentNew, String token) throws NoSuchElementException, AuthenticationException {
         Optional<Comment> commentOptional = commentRepository.findById(id);
         if (commentOptional.isEmpty()) throw new NoSuchElementException();
         Comment commentDB = commentOptional.get();
+        if (!AccountService.authenticateUser(commentDB.getWriter().getName(), token)) {
+            throw new AuthenticationException("Invalid token");
+        }
         commentDB.setText(commentNew.getText());
         return commentRepository.save(commentDB).getCommentId();
     }
 
-    public Long deleteComment(Long id) throws NoSuchElementException {
+    public Long deleteComment(Long id, String token) throws NoSuchElementException, AuthenticationException {
         Optional<Comment> commentOptional = commentRepository.findById(id);
         if (commentOptional.isEmpty()) throw new NoSuchElementException();
         Comment commentDB = commentOptional.get();
+        if (!AccountService.authenticateUser(commentDB.getWriter().getName(), token)) {
+            throw new AuthenticationException("Invalid token");
+        }
         commentRepository.deleteById(id);
         return commentDB.getCommentId();
     }
