+++ backend/src/main/java/de/damien/funnyplaces/places/PlaceService.java
@@ -1,11 +1,13 @@
 package de.damien.funnyplaces.places;
 
+import de.damien.funnyplaces.accounts.AccountService;
 import de.damien.funnyplaces.comments.Comment;
 import de.damien.funnyplaces.comments.CommentRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import javax.naming.AuthenticationException;
 import java.util.List;
 import java.util.NoSuchElementException;
 import java.util.Optional;
@@ -22,30 +24,43 @@ public class PlaceService {
         this.commentRepository = commentRepository;
     }
 
-    public Place addPlace(Place place) {
+    public Place addPlace(Place place, String token) throws AuthenticationException {
+        if (!AccountService.authenticateUser(place.getCreator().getName(), token)) {
+            throw new AuthenticationException("Invalid token");
+        }
         return placeRepository.save(place);
     }
 
 
-    public Place getPlace(Long id) throws NoSuchElementException {
+    public Place getPlace(Long id, String token) throws NoSuchElementException, AuthenticationException {
         Optional<Place> placeOptional = placeRepository.findById(id);
         if (placeOptional.isEmpty()) {
             throw new NoSuchElementException();
         }
-        return placeOptional.get();
+        Place place = placeOptional.get();
+        if (AccountService.getAccountByToken(token) == null) {
+            throw new AuthenticationException("Invalid token");
+        }
+        return place;
     }
 
-    public List<Comment> getAllComments(Long id) {
+    public List<Comment> getAllComments(Long id, String token) throws AuthenticationException {
         Optional<Place> placeOptional = placeRepository.findById(id);
         if (placeOptional.isEmpty()) throw new NoSuchElementException();
         Place place = placeOptional.get();
+        if (AccountService.getAccountByToken(token) == null) {
+            throw new AuthenticationException("Invalid token");
+        }
         return place.getComments();
     }
 
-    public Place updatePlace(Long id, Place placeNew) throws NoSuchElementException {
+    public Place updatePlace(Long id, Place placeNew, String token) throws NoSuchElementException, AuthenticationException {
         Optional<Place> place = placeRepository.findById(id);
         if (place.isEmpty()) throw new NoSuchElementException();
         Place placeDB = place.get();
+        if (!AccountService.authenticateUser(placeDB.getCreator().getName(), token)) {
+            throw new AuthenticationException("Invalid token");
+        }
         if (placeNew.getTitle() != null && !placeNew.getTitle().isBlank())
             placeDB.setTitle(placeNew.getTitle());
         if (placeNew.getDescription() != null && !placeNew.getDescription().isBlank())
@@ -55,10 +70,13 @@ public class PlaceService {
     }
 
     @Transactional
-    public Place deletePlace(Long id) throws NoSuchElementException {
+    public Place deletePlace(Long id, String token) throws NoSuchElementException, AuthenticationException {
         Optional<Place> placeOptional = placeRepository.findById(id);
         if (placeOptional.isEmpty()) throw new NoSuchElementException();
         Place place = placeOptional.get();
+        if (!AccountService.authenticateUser(place.getCreator().getName(), token)) {
+            throw new AuthenticationException("Invalid token");
+        }
         deleteAllComments(place.getComments());
         placeRepository.deleteById(id);
         return place;
@@ -72,7 +90,10 @@ public class PlaceService {
     }
 
     @Transactional
-    public List<Place> getAllPlaces() {
+    public List<Place> getAllPlaces(String token) throws AuthenticationException {
+        if (AccountService.getAccountByToken(token) == null) {
+            throw new AuthenticationException("Invalid token");
+        }
         return placeRepository.findAll();
     }
 }
